+++ src/Npgsql/NpgsqlExecute.cs
@@ -27,8 +27,6 @@
 
 using System;
 using System.IO;
-using System.Text;
-using System.Net;
 
 namespace Npgsql
 {
